DB/requirements.txt
PyQt5
psycopg2-binary
python-dotenv
